excess treatment

+++ src/films/films.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, MethodNotAllowedException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm'
 import { FilmsEntities } from './entities/filmesEntities';
@@ -15,16 +15,22 @@ export class FilmsService {
 
     async assessmentByUser(filmsDTO : responseFilmsDTO):Promise<FilmsEntities>{
         const films = this.filmsRepository.create(filmsDTO)
+        if(!films){
+            throw new ForbiddenException("Enter a valid movie")
+        }
          
         return this.filmsRepository.save(films);
     }
 
 
     async updateFilms(updateFilms:updateFilmsDTO , filmId:number):Promise<FilmsEntities>{
+        if(!filmId){
+            throw new MethodNotAllowedException('Enter a valid id')
+        }
         const film = await this.filmsRepository.findOne({ where: { id: filmId } });
 
         if(!film){
-            throw new Error("Filme não encontrado")
+            throw new NotFoundException("Film not found")
         }   
 
         return this.filmsRepository.save(
@@ -37,10 +43,13 @@ export class FilmsService {
 
 
     async deleteFilms(filmId: number):Promise<void>{
+        if(!filmId){
+            throw new MethodNotAllowedException('Enter a valid id')
+        }
         const film = await this.filmsRepository.findOne({where:{id:filmId}})
 
         if(!film){
-            throw new Error("Filme não encontrado")
+          throw new NotFoundException("Film not found")
         }
 
         await this.filmsRepository.delete(film)
